fix: 🐛 moved-app-types-to-the-root
Moved @firebase/app-types declaration overrides to the root.
